[TASK] Ensure jobrouter_base extension is required in function tests
Functional tests requires root extension and other extensions. However,
the `jobrouter_base` extension is not required in the function tests.
Because of this fact, that extension is not symlinked by the TYPO3
testing-framework to the functional test folders.

This change requires the extensions as `test extensions to load` to
ensure the prescence of the base extension during the functional tests.